refactor Ability
