+++ src/assets/Hero.java
@@ -81,7 +81,6 @@ public class Hero extends Monster {
 			int numTar = this.getNumTar();
 			action = true;
 
-			System.out.println("possible targets: " + possTargets.size());
 			if (numTar == -1 || numTar >= possTargets.size())
 				this.addTargets(possTargets);
 			else {
