+++ src/combat/moves/magic/Possess.java
@@ -1,6 +1,6 @@
 package combat.moves.magic;
 
-import combat.Ability;
+import combat.moves.Ability;
 import combat.Status;
 import assets.Monster;
 import main.Interface;
