+++ src/combat/moves/magic/Shock.java
@@ -1,7 +1,8 @@
 package combat.moves.magic;
 
 import assets.*;
-import combat.*;
+import combat.moves.Ability;
+import combat.Status;
 import main.Interface;
 
 public class Shock extends Ability {
