+++ src/combat/moves/melee/ChargeAttack.java
@@ -1,7 +1,8 @@
 package combat.moves.melee;
 
 import assets.*;
-import combat.*;
+import combat.moves.Ability;
+import combat.Status;
 import main.Interface;
 
 public class ChargeAttack extends Ability {
